refactor(spec): move observability seam into the spec layer
The spec layer's last upward dependency on node was the observability
import in fork_choice.py and state_transition.py. A prior refactor moved
the Interval type and protocol constants into the spec layer but left
this edge, so spec still reached up into node.

Relocate node/observability/ to spec/observability/. The package is
purely the vendor-neutral observer seam: a SpecObserver Protocol, a
no-op default, the observe_* context-manager timers, and set_observer.
Spec emits through it with a no-op default, so spec imports stay
side-effect-free. The concrete Prometheus backend stays in node/metrics
and the startup set_observer wiring stays in the CLI; only the seam moves.

Pure relocation, no behavior change. The test moves to the mirrored
spec path per the test-structure convention.

After this, grep -rn "lean_spec.node" src/lean_spec/spec/ is empty: the
spec layer no longer depends on node in either direction.

Alternative considered: leave the seam in node and inject an observer
into every spec entry point. Rejected as strictly worse churn that
fights the existing module-level singleton design; the seam is itself a
spec-level concern (the spec defines what it observes), so it belongs in
spec with the no-op default.
