+++ src/lean_spec/spec/observability/observer.py
